Enhance SaleReturnService and SaleService for item processing and initialization
Update SaleReturnService to include more descriptive error messages and ensure items are initialized as an empty list if null. Modify SaleService similarly to handle item initialization. Adjust PurchaseOrderData to change supplierId type from Integer to Long for consistency. Refactor BaseIntegrationTest to extend from SigrapBackendApplication and streamline test configurations. Update DataSeederTest to improve clarity and remove unnecessary comments. Enhance AttendanceIntegrationTest with CRUD operations and validation constraints for attendance records.

+++ src/main/java/com/sigrap/supplier/PurchaseOrderData.java
@@ -31,7 +31,7 @@ import lombok.NoArgsConstructor;
  * <p>Usage Example:
  * <pre>
  * PurchaseOrderData orderData = PurchaseOrderData.builder()
- *     .supplierId(1)
+ *     .supplierId(1L)
  *     .deliveryDate(LocalDate.now().plusWeeks(1))
  *     .items(List.of(orderItem1, orderItem2))
  *     .build();
@@ -61,7 +61,7 @@ public class PurchaseOrderData {
    */
   @NotNull(message = "Supplier ID cannot be null")
   @Schema(description = "ID of the supplier for this order", example = "1")
-  private Integer supplierId;
+  private Long supplierId;
 
   /**
    * The delivery date for the order.
